+++ kubernetes-popups.el
@@ -5,6 +5,7 @@
 (require 'magit-popup)
 (require 'transient)
 (require 'kubernetes-contexts)
+(require 'kubernetes-process)
 (require 'kubernetes-state)
 (require 'kubernetes-utils)
 
@@ -27,6 +28,23 @@
 
 ;; Popup definitions
 
+(transient-define-prefix kubernetes-proxy ()
+  [["Connection"
+    ("=p" "Port" "--port=" read-string)]]
+  [["Actions"
+    ;; TODO: Update this label dynamically based on current proxy status
+    ("P" "Enable/disable" kubernetes-proxy-toggle)]])
+
+(defun kubernetes-proxy-toggle (enable-disable args)
+  "Enable/disable kubectl proxy according to ENABLE-DISABLE, using ARGS.
+
+If disabling the proxy, ARGS is ignored."
+  (interactive (list (not (proxy-active-p kubernetes--global-process-ledger))
+                     (transient-args 'kubernetes-proxy)))
+  (if enable-disable
+      (get-proxy-process kubernetes--global-process-ledger args)
+    (kill-proxy-process kubernetes--global-process-ledger)))
+
 (transient-define-prefix kubernetes-dispatch ()
   [["Environment"
     ("c" "Configuration" kubernetes-config-popup)]
@@ -40,7 +58,8 @@
     ("e" "Exec" kubernetes-exec)
     ("f" "File" kubernetes-file)
     ("L" "Labels" kubernetes-labels)
-    ("l" "Logs" kubernetes-logs)]])
+    ("l" "Logs" kubernetes-logs)
+    ("P" "Proxy" kubernetes-proxy)]])
 
 (transient-define-prefix kubernetes-exec ()
   "Execute into Kubernetes resources."