+++ tests/test-process.el
@@ -36,6 +36,102 @@
                ((eq proc :fake-proc-live) t)
                ((eq proc :fake-proc-dead) nil)))))
 
+  (describe "kill-proxy-process"
+    (before-each
+      (spy-on 'kubernetes-process-kill-quietly))
+    (it "does nothing if no proxy process present"
+      (kill-proxy-process (kubernetes--process-ledger :proxy nil))
+      (expect 'kubernetes-process-kill-quietly :not :to-have-been-called))
+    (it "terminates and wipes existing proxy process"
+      (setq ledger (kubernetes--process-ledger
+                    :proxy (kubernetes--ported-process-record
+                            :process :existing-process
+                            :address "localhost"
+                            :port 9999)))
+      (kill-proxy-process ledger)
+      (expect 'kubernetes-process-kill-quietly :to-have-been-called)
+      (expect (oref ledger proxy) :to-equal nil)))
+
+  (describe "proxy-active-p"
+    (it "returns nil if there is no record"
+      (expect (proxy-active-p (kubernetes--process-ledger :proxy nil))
+              :to-equal
+              nil))
+    (it "returns nil if record has nil process"
+      (expect (proxy-active-p (kubernetes--process-ledger
+                               :proxy (kubernetes--ported-process-record
+                                       :process nil)))
+              :to-equal
+              nil))
+    (it "returns nil if record's process is not alive"
+      (spy-on 'process-live-p :and-return-value nil)
+      (expect (proxy-active-p (kubernetes--process-ledger
+                               :proxy (kubernetes--ported-process-record
+                                       :process :fake-process)))
+              :to-equal
+              nil)
+      (expect 'process-live-p :to-have-been-called))
+    (it "returns t if record's process is alive"
+      (spy-on 'process-live-p :and-return-value t)
+      (expect (proxy-active-p (kubernetes--process-ledger
+                               :proxy (kubernetes--ported-process-record
+                                       :process :fake-process)))
+              :to-equal
+              t)
+      (expect 'process-live-p :to-have-been-called)))
+
+  (describe "get-proxy-process"
+    (before-each
+      (spy-on 'kill-proxy-process)
+      (spy-on 'proxy-ready-p :and-return-value t)
+      (spy-on 'kubernetes-kubectl :and-return-value :new-kubectl-proc)
+      (spy-on 'sleep-for))
+    (describe "when proxy present in ledger"
+      (describe "when port specified in arguments"
+        :var (args)
+        (describe "when argument port doesn't match port of existing proxy"
+          (before-each
+            (setq args '("--port=1111"))
+            (setq ledger (kubernetes--process-ledger
+                          :proxy (kubernetes--ported-process-record
+                                  :port 9999))))
+          (it "destroys and creates new proxy process"
+            (get-proxy-process ledger args)
+            (expect 'kill-proxy-process :to-have-been-called)
+            (expect (oref (oref ledger proxy) port) :to-equal 1111)))
+        (describe "when argument port matches port of existing proxy"
+          (before-each
+            (setq args '("--port=1111"))
+            (setq ledger (kubernetes--process-ledger
+                          :proxy (kubernetes--ported-process-record
+                                  :process :fake-process
+                                  :port 1111))))
+          (it "returns the existing proxy process"
+            (expect (get-proxy-process ledger args) :to-equal :fake-process)
+            (expect 'kill-proxy-process :not :to-have-been-called)
+            (expect (oref (oref ledger proxy) port) :to-equal 1111))))
+      (describe "when no port specified in arguments"
+        (before-each
+          (setq ledger (kubernetes--process-ledger
+                        :proxy (kubernetes--ported-process-record
+                                :process :fake-process))))
+        (it "returns the existing proxy process"
+          (expect (get-proxy-process ledger '("--something=else"))
+                  :to-equal
+                  :fake-process))))
+    (describe "when proxy not present in ledger"
+      (before-each
+        (setq ledger (kubernetes--process-ledger :proxy nil)))
+      (describe "when no port specified in arguments"
+        (it "creates new proxy process with default port"
+          (let ((kubernetes-default-proxy-port 1234))
+            (expect (get-proxy-process ledger nil) :to-equal :new-kubectl-proc)
+            (expect (oref (oref ledger proxy) port) :to-equal 1234))))
+      (describe "when port specified in arguments"
+        (it "creates new proxy process with that port"
+          (get-proxy-process ledger '("--port" "1111"))
+          (expect (oref (oref ledger proxy) port) :to-equal 1111)))))
+
   (describe "process lookup"
     (it "returns the process for resources"
       (setq ledger (kubernetes--process-ledger