+++ docs/CHANGELOG.md
@@ -32,9 +32,13 @@ versioning][semver].
     (truncated) view of the data.
     ([#245](https://github.com/kubernetes-el/kubernetes-el/pull/245))
 -   Added mapping for E to enable entity editing in evil mode.
-    ([#246](https://github.com/kubernetes-el/kubernetes-el/pull/246))
+    ([#246](https://github.com/kubernetes-el/kubernetes-el/pull/246)).
+-   Added ability to enable and disable [`kubectl` proxies][kubectl proxy] via
+    `P P`, with status display in the overview. ([#252](https://github.com/kubernetes-el/kubernetes-el/pull/252))
 -   `kubernetes-overview` now returns an error if `kubectl` or
     `kubernetes-kubectl-executable` is not found on PATH.
+
+[kubectl proxy]: https://kubernetes.io/docs/tasks/extend-kubernetes/http-proxy-access-api/
 
 ### Refinements